Revert "changes nahj admin"

+++ src/app/pages/nahj/home/home.component.ts
@@ -2,8 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { GeoService } from '../../../services/geo/geo.service';
 import { EvalutionStatusService } from '../../../services/evaluationStatus/evalution-status.service';
 import { StudyYearsService } from '../../../services/studyYears/study-years.service';
-import { ConfigService } from '../../../services/config';
-import { CityService } from '../../../services/city/city.service';
+import {map} from 'rxjs/operators';
 
 @Component({
   selector: 'app-home',
@@ -12,7 +11,7 @@ import { CityService } from '../../../services/city/city.service';
 })
 export class HomeComponent implements OnInit {
 
-	url : string;
+	url : string = "http://localhost:4466";
 	geoArray : any = [];
 	citiesArray : any = [];
 	termArray = [];
@@ -25,13 +24,8 @@ export class HomeComponent implements OnInit {
 
 	constructor(
 		private geoService: GeoService,
-		private cityService: CityService,
 		private evaluationService: EvalutionStatusService,
-		private studyYearService: StudyYearsService,
-		private configService: ConfigService) {
-
-			this.url = this.configService.url;
-		}
+		private studyYearService: StudyYearsService) { }
 
  
 	ngOnInit() {
@@ -43,31 +37,25 @@ export class HomeComponent implements OnInit {
 
 	//get data functions
 
-	getGeoCityData(id){//get geoDataCity
+	getGeoCityData(name){//get geoDataCity
 		this.geoService.service({
 			method: "GET",
 			url: this.url
 		}).subscribe((data: any) => {
-			console.log("get geo city", data)
-			if (data.data) {
-				this.geoArray = data.data.geoAreas.map((item, index) => {
-					if(!name && index == 0){
-						this.citiesArray = item["cities"]
-					}else if(item.id == id){
-						this.citiesArray =  item["cities"]
-					}
-					this.selectedCity = undefined;
-					return item;
-				} );
-			}
+			this.geoArray = data.data.geoAreas.map((item, index) => {
+				if(!name && index == 0){
+					this.citiesArray = item["cities"]
+				}else if(item.name == name){
+					this.citiesArray =  item["cities"]
+				}
+				return item.name;
+			} );
 		});
 	}
 
 	getCities($event){
-		console.log("list", $event)
-		this.getGeoCityData($event.value.id || undefined);
-		this.selectedGeo = $event.value;
-		this.selectedCity = undefined;		
+		this.getGeoCityData($event.value || undefined);
+		this.selectedGeo = $event.value 
 	}
 	getStudyYears(){
 		this.studyYearService.service({
@@ -75,7 +63,7 @@ export class HomeComponent implements OnInit {
 			url: this.url
 		}).subscribe(data => {
 			this.termArray = data["data"].studyYears.map(styYear =>{
-				return styYear;
+				return styYear.name;
 			})
 
 		})
@@ -86,7 +74,7 @@ export class HomeComponent implements OnInit {
 			method: "GET",
 			url: this.url
 		}).subscribe((data: any) => {
-			this.evaluationStatus = data.data.evaluationStatuses;
+			this.evaluationStatus = data.data.evaluationStatuses.map(item => item.name);
 		})
 	}
 
@@ -100,53 +88,115 @@ export class HomeComponent implements OnInit {
 					name: $event.newValue
 				}).subscribe((data: any) => {
 					this.getGeoCityData($event.newValue);
-					// this.selectedGeo = $event.newValue;
-				});			
+					this.selectedGeo = $event.newValue;
+				});
+			
 			break;
 			case "update" :
 				this.geoService.service({
 					url:this.url,
-					method: 'POST',
-					newName:$event.newValue,
-					name:$event.value,
-					id:this.selectedGeo['id']
+					method: 'GET',
 				}).subscribe((data:any)=>{
-					this.getGeoCityData($event.newValue)
+					data.data.geoAreas.map((item)=>{
+						if(item.name === $event.value){
+							this.geoService.service({
+								url:this.url,
+								method: 'POST',
+								newName:$event.newValue,
+								name:$event.value,
+								cities:item.cities
+							}).subscribe((data:any)=>{
+								this.getGeoCityData($event.newValue)
+								this.geoService.service({
+									method: 'DELETE',
+									url: this.url,
+									name: $event.value
+								}).subscribe((data: any)=>{
+									this.getGeoCityData($event.newValue)
+								})
+							})					
+						}
+					})
 				})
 			break;
 		}
 	}
 
 	addNewCity($event){
+		console.log("add city", $event)
 		switch ($event.eventType) {
 			case "add":
-				this.cityService.service({
-					method: "PUT",
+				this.geoService.service({
 					url: this.url,
-					name: $event.newValue,
-					id: this.selectedGeo['id']
-				}).subscribe(cities => {
-					console.log("cities", cities)
-					this.getGeoCityData(this.selectedGeo['id'])
-				})
-			break;
+					name: this.selectedGeo
+				}).subscribe(data => {
+					let oldCities = data["data"]["geoAreas"][0]["cities"];
+					let newCities = oldCities
+					console.log("old city", oldCities)
+					newCities.push($event.newValue);
+					console.log("new city", newCities)
+
+					this.geoService.service({
+						method: 'GET',
+						url: this.url
+					}).subscribe(geos=>{
+						geos['data'].geoAreas.filter(geo=>geo.name == this.selectedGeo).map(geo=>{
+							this.geoService.service({
+								method: "DELETE",
+								url: this.url,
+								Id: geo.id
+							}).subscribe(data =>{
+								this.geoService.service({
+									method: 'PUT',
+									url: this.url,
+									name: this.selectedGeo,
+									cities: newCities
+								}).subscribe((data: any) => {
+									this.citiesArray= data["data"]["createGeoArea"]["cities"];
+									this.selectedCity = $event.newValue;
+								});
+							})
+						})
+					})
+				});
+				break;
 
 			case "update":
-				this.cityService.service({
-					method: "POST",
-
+				this.geoService.service({
 					url: this.url,
-					name: $event.newValue,
-					id: this.selectedCity['id']
-				}).subscribe(cities => {
-					this.getGeoCityData(this.selectedGeo['id'])
-				})
-			break;
+					name: this.selectedGeo
+				}).subscribe(data => {
+					let cities = data["data"]["geoAreas"][0]["cities"];
+					let index = cities.indexOf($event.value);
+					cities.splice(index , 1);
+					cities.push($event.newValue);
+					this.geoService.service({
+						method: 'GET',
+						url: this.url
+					}).subscribe(geos=>{
+						geos['data'].geoAreas.filter(geo=>geo.name == this.selectedGeo).map(geo=>{
+							this.geoService.service({
+								method: "DELETE",
+								url: this.url,
+								Id: geo.id
+							}).subscribe(data =>{
+								this.geoService.service({
+									method: 'PUT',
+									url: this.url,
+									name: this.selectedGeo,
+									cities: cities
+								}).subscribe((data: any) => {
+									this.citiesArray= data["data"]["createGeoArea"]["cities"];
+									this.selectedCity = $event.newValue;
+								});
+							})
+						})
+					})
+				});
+				break;
 		}
 	}
 	addNewEvaluation($event){
-		console.log("add eval", $event)
-		console.log("selected eval", this.selectedEval)
 		switch ($event.eventType) {
 			case "add":
 				this.evaluationService.service({
@@ -155,19 +205,28 @@ export class HomeComponent implements OnInit {
 					name: $event.newValue
 				}).subscribe((data: any) => {
 					this.getEvaluationStatusData();
+					this.selectedEval = $event.newValue;
 				});
-			break;
+				break;
 		
 			case "update":
 				this.evaluationService.service({
-					method: "POST",
-					url: this.url,
-					name: $event.newValue,
-					id: this.selectedEval['id']
+					method: "GET",
+					url: this.url
 				}).subscribe((data: any) => {
-					this.getEvaluationStatusData();
-				});
-			break;
+					data.data.evaluationStatuses.filter(item => item.name == $event.value ).map(item => {
+						this.evaluationService.service({
+							method: "POST",
+							url: this.url,
+							name: $event.newValue,
+							id: item.id
+						}).subscribe((data: any) => {
+							this.getEvaluationStatusData();
+							this.selectedEval = $event.newValue;
+						});
+					});
+				})
+				break;
 		}
 	}
 	addNewStudyYear($event){
@@ -179,17 +238,26 @@ export class HomeComponent implements OnInit {
 					name: $event.newValue, 
 				}).subscribe(data=>{
 					this.getStudyYears();
+					this.selectedTerm  =  $event.newValue;
 				});
 			break;
 			case "update":
-				this.studyYearService.service({
-					method: "POST",
-					url: this.url,
-					name: $event.newValue,
-					Id: this.selectedTerm['id'] 
-				}).subscribe(data=>{
-					this.getStudyYears();
-				});
+			this.studyYearService.service({
+				method: "GET",
+				url: this.url 
+			}).subscribe(data=>{
+				data['data'].studyYears.filter(studyYear=> studyYear.name == $event.value).map(resp=>{
+					this.studyYearService.service({
+						method: "POST",
+						url: this.url,
+						name: $event.newValue,
+						Id: resp.id 
+					}).subscribe(data=>{
+						this.getStudyYears();
+						this.selectedTerm  =  $event.newValue;
+					});
+				})
+			});
 			break;
 		}
 	}
@@ -197,24 +265,50 @@ export class HomeComponent implements OnInit {
 	//delete data functions
 	deleteGeo($event){
 		this.geoService.service({
-			method: 'DELETE',
-			url: this.url,
-			id: this.selectedGeo['id']
-		}).subscribe((data:any)=> {
-			this.getGeoCityData(undefined);
-			this.selectedGeo = undefined;
-		})	
+			method: 'GET',
+			url: this.url
+		}).subscribe(geos=>{
+			geos['data'].geoAreas.map(geo=>{
+				if (geo.name == $event.value) {
+					this.geoService.service({
+						method: 'DELETE',
+						url: this.url,
+						Id: geo.id
+					}).subscribe((data:any)=> {
+						this.getGeoCityData(undefined);
+						this.selectedGeo = undefined;
+					})
+				}
+			})
+		})
 	}
 
 	deleteCity($event){
-		this.cityService.service({
-			method: 'DELETE',
-			url: this.url,
-			id: this.selectedCity['id']
-		}).subscribe((data:any)=> {
-			this.getGeoCityData(this.selectedGeo['id']);
-			this.selectedCity = undefined;
-		})	
+		this.geoService.service({
+			method: 'GET',
+			url: this.url
+		}).subscribe(data=>{
+			data['data'].geoAreas.filter(geo=>geo.name == this.selectedGeo).map(geo=>{
+				let cities = geo["cities"];
+				let index = cities.indexOf($event.value);
+				cities.splice(index , 1);
+				this.geoService.service({
+					method: "DELETE",
+					url: this.url,
+					Id: geo.id
+				}).subscribe(data =>{
+					this.geoService.service({
+						method: 'PUT',
+						url: this.url,
+						name: this.selectedGeo,
+						cities: cities
+					}).subscribe((data: any) => {
+						this.citiesArray= data["data"]["createGeoArea"]["cities"];
+						this.selectedCity = $event.newValue;
+					});
+				})
+			})
+		})
 	}
 
 	deleteEvaluation($event){
@@ -252,5 +346,12 @@ export class HomeComponent implements OnInit {
 				});
 			})
 		});
+		// this.studyYearService.service({
+		// 	method: "DELETE",
+		// 	url: this.url
+		// }).subscribe(data =>{
+		// 	this.termArray = []
+		// 	this.selectedTerm = undefined;
+		// })
 	}
 }
